Update bmptoylib.c
I've decided not to finish adding drawing functions to the library. It is supposed to be as simple as possible and programmers with more skill than me can implement better drawing functions anyway. The library gives beginner C programmers an easy to use library for the BMP file format which they can play around with.